+++ chats/apps/api/v1/msgs/viewsets.py
@@ -1,11 +1,15 @@
 from django.db import transaction
 from django_filters.rest_framework import DjangoFilterBackend
+from django.core.exceptions import ObjectDoesNotExist
 from pydub.exceptions import CouldntDecodeError
-from rest_framework import filters, mixins, pagination, parsers, status, viewsets
+from rest_framework import filters, mixins, parsers, status, viewsets
 from rest_framework.decorators import action
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
+from chats.apps.api.pagination import (
+    CustomCursorPagination,
+)
 from chats.apps.api.v1.msgs.filters import MessageFilter, MessageMediaFilter
 from chats.apps.api.v1.msgs.permissions import MessageMediaPermission, MessagePermission
 from chats.apps.api.v1.msgs.serializers import (
@@ -32,6 +36,9 @@ class MessageViewset(
     permission_classes = [IsAuthenticated, MessagePermission]
     lookup_field = "uuid"
 
+    pagination_class = CustomCursorPagination
+    ordering = ["-created_on"]
+
     def get_paginated_response(self, data):
         if self.request.query_params.get("reverse_results", False):
             data.reverse()
@@ -47,6 +54,23 @@ class MessageViewset(
         with transaction.atomic():
             serializer.save()
             serializer.instance.notify_room("create", True)
+
+            message = serializer.instance
+            if message.user and message.room.first_user_assigned_at:
+                try:
+                    metric = message.room.metric
+                    if metric.first_response_time is None:
+                        from chats.apps.dashboard.tasks import (
+                            calculate_first_response_time_task,
+                        )
+
+                        calculate_first_response_time_task.delay(str(message.room.uuid))
+                except ObjectDoesNotExist:
+                    from chats.apps.dashboard.tasks import (
+                        calculate_first_response_time_task,
+                    )
+
+                    calculate_first_response_time_task.delay(str(message.room.uuid))
 
     def perform_update(self, serializer):
         serializer.save()
@@ -88,8 +112,10 @@ class MessageMediaViewset(
     filterset_class = MessageMediaFilter
     parser_classes = [parsers.MultiPartParser]
     permission_classes = [IsAuthenticated, MessageMediaPermission]
-    pagination_class = pagination.PageNumberPagination
+    pagination_class = CustomCursorPagination
     lookup_field = "uuid"
+    ordering = "created_on"
+    ordering_fields = ["created_on", "content_type"]
 
     def get_queryset(self):
         if self.request.query_params.get("contact") or self.request.query_params.get(