+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 3.27.0
+# Add
+  - flag to indicate if contact has history
+
+# 3.26.7
+# Add
+  - disconnect agents with custom breaks
+
+# 3.26.6
+# Add
+  - Cursor pagination on messages list endpoint
+
 # 3.26.5
 # Add
   - Internal note feature
@@ -755,6 +767,8 @@
 
 ## Change
   - Function is attending now uses pendulum to parse timezone in sector.
+
+